requirements.txt
langchain-google-genai==0.0.10
python-dotenv==1.0.1
rdflib==7.0.0
streamlit==1.31.0
pandas==2.1.4
matplotlib==3.8.2 